+++ packages/grafana-data/src/geo/layer.ts
@@ -2,42 +2,13 @@ import { PluggableMap } from 'ol';
 import BaseLayer from 'ol/layer/Base';
 import { ReactNode } from 'react';
 
+import { MapLayerOptions as MapLayerOptionsSchema } from '@grafana/schema';
+
 import { EventBus } from '../events';
 import { GrafanaTheme2 } from '../themes';
 import { MatcherConfig, PanelData } from '../types';
 import { PanelOptionsEditorBuilder } from '../utils';
 import { RegistryItemWithOptions } from '../utils/Registry';
-
-/**
- * @alpha
- */
-export enum FrameGeometrySourceMode {
-  Auto = 'auto', // Will scan fields and find best match
-  Geohash = 'geohash',
-  Coords = 'coords', // lon field, lat field
-  Lookup = 'lookup', // keys > location
-  // H3 = 'h3',
-  // WKT = 'wkt,
-  // geojson? geometry text
-}
-
-/**
- * @alpha
- */
-export interface FrameGeometrySource {
-  mode: FrameGeometrySourceMode;
-
-  // Field mappings
-  geohash?: string;
-  latitude?: string;
-  longitude?: string;
-  h3?: string;
-  wkt?: string;
-  lookup?: string;
-
-  // Path to Gazetteer
-  gazetteer?: string;
-}
 
 /**
  * This gets saved in panel json
@@ -49,26 +20,10 @@ export interface FrameGeometrySource {
  *
  * @alpha
  */
-export interface MapLayerOptions<TConfig = any> {
-  type: string;
-  name: string; // configured unique display name
-
+export interface MapLayerOptions<TConfig = any> extends Omit<MapLayerOptionsSchema, 'config' | 'filterData'> {
   // Custom options depending on the type
   config?: TConfig;
-
-  // Common method to define geometry fields
-  location?: FrameGeometrySource;
-
-  // Defines which data query refId is associated with the layer
   filterData?: MatcherConfig;
-
-  // Common properties:
-  // https://openlayers.org/en/latest/apidoc/module-ol_layer_Base-BaseLayer.html
-  // Layer opacity (0-1)
-  opacity?: number;
-
-  // Check tooltip (defaults to true)
-  tooltip?: boolean;
 }
 
 /**
